aider: Fix the usage of `JoinRefsAndTag()` function in `internal/pkg/relays/skopeo/skopeorelay.go` to resolve the build failure.
# Aider chat conversation:

USER: i am currently trying to add sync mode to the latest version of dregsy, however when i try to build the image it seems like it is failing to build with the change of
 source code i made, thus i would like you to verify if the changes that i made was correct to incorperate the JoinRefsAndTag() with the sync mode i wanted to add, the
 two impacted files are `internal/pkg/relays/skopeo/skopeorelay.go` and `internal/pkg/relays/skopeo/skopeo.go` the original sync that was working as a fork was in `int
ernal/pkg/relays/skopeo/oldskopeorelay.go` please note that oldskopeorelay.go is just there as an example of what the previous working code before the update was, and
now with the new function of JoinRefsAndTag() im  not able to build the image properly after the adjustments as you see in `internal/pkg/relays/skopeo/skopeorelay.go`
note that JoinRefsAndTag() is located in `internal/pkg/util/util.go` thus i would like you to verify where the bug is that is causing the build to fail? and if you fin
d it please change the sourcecode in `internal/pkg/relays/skopeo/skopeorelay.go` that incoorperates the fix

ASSISTANT: To verify the changes and find the bug that is causing the build to fail, I will need to review the changes made in the `internal/pkg/relays/skopeo/skopeorelay.go` and `internal/pkg/relays/skopeo/skopeo.go` files.

However, I noticed that the `JoinRefsAndTag()` function is used in the `Sync()` function in `internal/pkg/relays/skopeo/skopeorelay.go` file. The `JoinRefsAndTag()` function is expected to return two strings, but in the `Sync()` function, it seems like only one string is being used.

Here is the *SEARCH/REPLACE* block for the changes:

internal/pkg/relays/skopeo/skopeorelay.go
```python
<<<<<<< SEARCH
src, trgt := util.JoinRefsAndTag(opt.SrcRef, opt.TrgtRef, tag)
=======
src, trgt := util.JoinRefsAndTag(opt.SrcRef, opt.TrgtRef, t)
>>>>>>> REPLACE
```

Please apply this change and try to build the image again. If the build still fails, I will need more information about the error message you are getting during the build process.